requirements.txt
streamlit==1.28.1
clickhouse-connect==0.6.23
openai==1.3.7
pandas==2.1.3
plotly==5.17.0
python-dotenv==1.0.0
pytz==2023.3
tabulate==0.9.0